+++ pkg/api/ssh/key/response.go
@@ -11,4 +11,24 @@ type (
 		} `json:"return"`
 		models.APIResponse
 	}
+	// GetResponse is the return from a GetRequest.
+	GetResponse struct {
+		Return models.SSHKey `json:"return"`
+		models.APIResponse
+	}
+	// AddResponse is the return from an AddRequest.
+	AddResponse struct {
+		Return struct {
+			KeyID string `json:"key_id"`
+		} `json:"return"`
+		models.APIResponse
+	}
+	// UpdateResponse is the return from an UpdateRequest.
+	UpdateResponse struct {
+		models.APIResponse
+	}
+	// RemoveResponse is the return from an RemoveRequest.
+	RemoveResponse struct {
+		models.APIResponse
+	}
 )