+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+
+## [v0.3.x] - 2032-03-15
+### Fixes
+- correct the ssh key update params
+
+
+## [v0.3.x] - 2032-03-13
+### Added
+- database list/get/update/remove for `/cloud/stack/image`
+- cloud stack image list_all `/cloud/image`
+- `/ssh/key/` add/update 
+
+## [v0.3.x] - 2023-03-07
+### Added
+- helper for handling boolish results
+- image list and get endpoints for `/cloud/image` endpoint
+
 ## [v0.3.2] - 2023-03-22
 ### Fixes
 - Fix a crash when unmarshalling when the `/server` returns a different type for the server disk size depending on the type of server.